+++ previous-versions/common/pkg/models/operations/getprofile.go
@@ -10,8 +10,6 @@ import (
 type GetProfileResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
-	ErrorMessage *shared.ErrorMessage
 	// OK
 	Profile *shared.Profile
 	// HTTP response status code for this operation
@@ -25,13 +23,6 @@ func (o *GetProfileResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *GetProfileResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *GetProfileResponse) GetProfile() *shared.Profile {